+++ tools/publish-modules.js
@@ -1,7 +1,7 @@
 const fs = require("node:fs");
 const path = require("node:path");
 const { parseArgs } = require("node:util");
-const runScript = require("@npmcli/run-script");
+const { execSync } = require("node:child_process");
 
 /** Core module. */
 const TARGET_CORE_MODULE = "core";
@@ -9,8 +9,14 @@ const TARGET_CORE_MODULE = "core";
 const TARGET_APPENDERS = "appenders";
 /** Project root directory. */
 const ROOT_DIR = path.join(__dirname, "..");
+/** Core module source directory. */
+const CORE_SRC_DIR = path.join(ROOT_DIR, "src/core");
+/** Appenders source directory. */
+const APPENDERS_SRC_DIR = path.join(ROOT_DIR, "src/appenders");
+/** Core module directory. */
+const CORE_DIST_DIR = path.join(ROOT_DIR, "dist/core");
 /** Appenders directory. */
-const APPENDERS_DIR = path.join(ROOT_DIR, "src/appenders");
+const APPENDERS_DIST_DIR = path.join(ROOT_DIR, "dist/appenders");
 
 /**
  * Publish a module. When module name is empty, publish core module.
@@ -22,10 +28,25 @@ async function publishModule(name) {
 
   // Publish.
   console.info(`Publish ${target}`);
-  return runScript({
-    event: `publish:${target}`,
-    path: ROOT_DIR,
-  });
+  if (target === TARGET_CORE_MODULE) {
+    execFileSync("vite", [
+      "build",
+      "--config",
+      path.join(CORE_SRC_DIR, "vite.config.js"),
+    ]);
+    return execFileSync("npm", ["publish", CORE_DIST_DIR, "--access=public"]);
+  } else {
+    execFileSync("vite", [
+      "build",
+      "--config",
+      path.join(APPENDERS_SRC_DIR, target, "vite.config.js"),
+    ]);
+    return execFileSync("npm", [
+      "publish",
+      path.join(APPENDERS_DIST_DIR, target),
+      "--access=public",
+    ]);
+  }
 }
 
 /**
@@ -46,8 +67,8 @@ function main() {
     // Publish core module.
     publishModule(TARGET_CORE_MODULE).then(() => {
       // Publish appenders.
-      const dirs = fs.readdirSync(APPENDERS_DIR).filter((file) => {
-        return fs.statSync(path.join(APPENDERS_DIR, file)).isDirectory();
+      const dirs = fs.readdirSync(APPENDERS_SRC_DIR).filter((file) => {
+        return fs.statSync(path.join(APPENDERS_SRC_DIR, file)).isDirectory();
       });
       dirs.forEach((dir) => {
         publishModule(dir);
@@ -55,8 +76,8 @@ function main() {
     });
   } else if (values.target === TARGET_APPENDERS) {
     // Publish appenders.
-    const dirs = fs.readdirSync(APPENDERS_DIR).filter((file) => {
-      return fs.statSync(path.join(APPENDERS_DIR, file)).isDirectory();
+    const dirs = fs.readdirSync(APPENDERS_SRC_DIR).filter((file) => {
+      return fs.statSync(path.join(APPENDERS_SRC_DIR, file)).isDirectory();
     });
     dirs.forEach((dir) => {
       publishModule(dir);